COMP90024-ClusterAndCloudComputing/Assignment2/backend/mastodon-harvester/app/requirements.txt
decorator==5.1.1
Mastodon.py==1.8.1
aiokafka==0.8.0
